Fix OverflowException in sync state and improve HTTP handling
- Add validation for negative values in ServerInfo.UpdateFolderSyncState
- Add PruneStaleFolderStates method to clean up stale folder states
- Add diagnostic event for invalid sync data detection
- Fix Connected Devices display showing -1 when no devices configured
- Improve HTTP request handling in SyncthingEventsListener and SyncthingMonitorService using HttpRequestMessage
- Fix MudBlazor 8.x warnings (Dense, Title, IsInitiallyExpanded attributes)
- Fix obsolete WatchSystemPreference warning in App.razor
- Remove unnecessary System.Text.Json package from WebUI
- Fix uninitialized field warnings in Profiling, Statistics, NavMenu

+++ src/CreatioHelper.Desktop/Models/SyncthingEvent.cs
@@ -106,13 +106,13 @@ public class FolderCompletionEventData
     /// Total global items
     /// </summary>
     [JsonPropertyName("globalItems")]
-    public int GlobalItems { get; set; }
+    public long GlobalItems { get; set; }
 
     /// <summary>
     /// Items still needed to sync
     /// </summary>
     [JsonPropertyName("needItems")]
-    public int NeedItems { get; set; }
+    public long NeedItems { get; set; }
 }
 
 /// <summary>
